packages/type-r-editor/pyproject.toml
[project]
name = "type-r-editor"
version = "0.1.0"
description = "Type-R text editor package."
readme = "README.md"
authors = [
    { name = "Wataru Shimoda", email = "wataru_shimoda@cyberagent.co.jp" },
]
requires-python = ">=3.10, <3.13"
dependencies = [
    "diffusers>=0.32.2",
    "easydict>=1.13",
    "einops>=0.8.1",
    "fsspec>=2024.2.0",
    "modelscope>=1.23.2",
    "numpy>=2.2.3",
    "open-clip-torch>=2.31.0",
    "pillow>=10.4.0",
    "pytorch-lightning>=2.5.0.post0",
    "safetensors>=0.5.3",
    "skia-python>=87.6",
    "torch",
    "transformers>=4.49.0",
    "ujson>=5.10.0",
    "type-r-core",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["type_r"]  # src layout does not work in editable mode.

[tool.uv.sources]
type-r-core = { workspace = true }